back/requirements.txt
asgiref==3.5.2
astroid==2.12.10
blis==0.7.8
catalogue==2.0.8
certifi
confection==0.0.3
cymem==2.0.6
dill==0.3.5.1
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
filelock==3.8.0
isort==5.10.1
Jinja2==3.1.2
langcodes==3.3.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
murmurhash==1.0.8
packaging==21.3
pathy==0.6.2
gunicorn
platformdirs==2.5.2
preshed==3.0.7
pydantic==1.9.2
pylint==2.15.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-decouple==3.6
pytz==2022.4
smart-open==5.2.1
sqlparse==0.4.3
srsly==2.4.4
thinc==8.1.3
tomli==2.0.1
tomlkit==0.11.5
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.4.0
wasabi==0.10.1
whitenoise==6.2.0
wrapt==1.14.1
